worked on buttons

+++ Summer-Excel/HomepageView.swift
@@ -16,12 +16,13 @@ class HomepageView: UIViewController {
     @IBOutlet weak var attendanceInput: UISwitch!
     
     @IBAction func pressEnter(_ sender: Any) {
-        let theseMiles = Int(mileInput.text!)
+        let theseMiles = Double(mileInput.text!)
         let theseMinutes = Int(minuteInput.text!)
         let theseNotes = notesInput.text
         let attend = attendanceInput.isOn
         
-        
+        //let thisTime = Time(min: Double)
+        //let thisWorkout = Workout(miles: theseMiles, timeE: Time, theDate: <#T##Date#>, words: <#T##String#>, attend: <#T##Bool#>)
     }
     
     }
